+++ ls_Whistles/core/changelog.lua
@@ -6,8 +6,15 @@ local _G = getfenv(0)
 
 -- Mine
 addon.CHANGELOG = [[
-### Loot Window
+### Action Bars
 
-- Added Transmog Loot Helper and Can I Mog It? support.
-- While not in combat you'll be able to use the interaction hotkey and/or space to loot everything.
+- Added a set of options to adjust the usable equipment highlight. It's the green highlight you see for usable trinkets and stuff like that. Can be found in the advanced options.
+
+### Adventure Guide
+
+- Fixed an issue where opening the suggested content frame while in combat could result in an error.
+
+### Character Frame
+
+- Fixed an issue where ilvl colouring wouldn't apply on the initial login.
 ]]