+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## Version 120001.02
+
+### Action Bars
+
+- Added a set of options to adjust the usable equipment highlight. It's the green highlight you see
+  for usable trinkets and stuff like that. Can be found in the advanced options.
+
+### Adventure Guide
+
+- Fixed an issue where opening the suggested content frame while in combat could result in an error.
+
+### Character Frame
+
+- Fixed an issue where ilvl colouring wouldn't apply on the initial login.
+
 ## Version 120001.01
 
 ### Loot Window